models: widen entity ids i32 -> i64 to match sea-orm 2.0 BIGINT PKs
The loco 0.16->1.1 / sea-orm 1.1->2.0 upgrade changed auto-increment
primary keys from INTEGER to BIGINT: loco's ColType::PkAuto now maps to
big_pk_auto, and its create_table foreign-key helper emits BigInteger to
match the referenced 64-bit PK. The generated entity structs, however,
still declared id/*_id as i32, so every query decoding an id against a
freshly-migrated Postgres failed with:

  Query Error: decoding column "id": Rust type Option<i32> (INT4) is not
  compatible with SQL type INT8

which surfaced as HTTP 500 on /api/auth/current the first time a tenant was
promoted onto a real Postgres, sending the SPA into a sign-in redirect loop.

Why it hid until now:
- the test suite runs on SQLite, which is loosely typed (INT4/INT8 are
  interchangeable), so the mismatch never tripped in CI;
- the deployment had no live tenant DBs (canaries route to a static backend,
  no castle app, no castle schema), so the sea-orm-2.0 migrations had never
  run against Postgres until the first promote.

Fix: widen every PK and FK to i64 across entities, model helpers, controller
Path/params, views, and test fixtures - exactly what loco 1.1's entity
generator produces today. No migration change needed (the columns are already
BIGINT); this only aligns the Rust side. Stale "INTEGER NOT NULL, FK" comments
in the migrations corrected to BIGINT.

Verified against a throwaway Postgres 18.6 (production's version): the full
request-test suite - which does INSERT ... RETURNING id and reads rows back
into the entities, the exact operation that 500'd - passes 11/11, where it
fails on the old i32 entities.

+++ src/controllers/findings.rs
@@ -75,14 +75,14 @@ fn validate_finding_text(
     Ok(())
 }
 
-async fn load_project(ctx: &AppContext, id: i32) -> Result<projects::Model> {
+async fn load_project(ctx: &AppContext, id: i64) -> Result<projects::Model> {
     projects::Entity::find_by_id(id)
         .one(&ctx.db)
         .await?
         .ok_or_else(|| Error::NotFound)
 }
 
-async fn load_finding(ctx: &AppContext, id: i32) -> Result<findings::Model> {
+async fn load_finding(ctx: &AppContext, id: i64) -> Result<findings::Model> {
     findings::Entity::find_by_id(id)
         .one(&ctx.db)
         .await?
@@ -98,7 +98,7 @@ fn is_privileged(user: &users::Model, membership: &Option<project_members::Model
 #[debug_handler]
 pub async fn create(
     CurrentUser(user): CurrentUser,
-    Path(project_id): Path<i32>,
+    Path(project_id): Path<i64>,
     State(ctx): State<AppContext>,
     Json(params): Json<CreateFindingParams>,
 ) -> Result<Response> {
@@ -166,7 +166,7 @@ pub async fn create(
 #[debug_handler]
 pub async fn list(
     CurrentUser(user): CurrentUser,
-    Path(project_id): Path<i32>,
+    Path(project_id): Path<i64>,
     State(ctx): State<AppContext>,
 ) -> Result<Response> {
     let project = load_project(&ctx, project_id).await?;
@@ -188,7 +188,7 @@ pub async fn list(
 #[debug_handler]
 pub async fn show(
     CurrentUser(user): CurrentUser,
-    Path(finding_id): Path<i32>,
+    Path(finding_id): Path<i64>,
     State(ctx): State<AppContext>,
 ) -> Result<Response> {
     let finding = load_finding(&ctx, finding_id).await?;
@@ -229,7 +229,7 @@ pub async fn show(
 #[debug_handler]
 pub async fn update(
     CurrentUser(user): CurrentUser,
-    Path(finding_id): Path<i32>,
+    Path(finding_id): Path<i64>,
     State(ctx): State<AppContext>,
     Json(params): Json<UpdateFindingParams>,
 ) -> Result<Response> {
@@ -302,7 +302,7 @@ pub async fn update(
 #[debug_handler]
 pub async fn publish(
     CurrentUser(user): CurrentUser,
-    Path(finding_id): Path<i32>,
+    Path(finding_id): Path<i64>,
     State(ctx): State<AppContext>,
 ) -> Result<Response> {
     let finding = load_finding(&ctx, finding_id).await?;
@@ -324,7 +324,7 @@ pub async fn publish(
 #[debug_handler]
 pub async fn unpublish(
     CurrentUser(user): CurrentUser,
-    Path(finding_id): Path<i32>,
+    Path(finding_id): Path<i64>,
     State(ctx): State<AppContext>,
 ) -> Result<Response> {
     let finding = load_finding(&ctx, finding_id).await?;
@@ -347,7 +347,7 @@ pub async fn unpublish(
 #[debug_handler]
 pub async fn remove(
     CurrentUser(user): CurrentUser,
-    Path(finding_id): Path<i32>,
+    Path(finding_id): Path<i64>,
     State(ctx): State<AppContext>,
 ) -> Result<Response> {
     let finding = load_finding(&ctx, finding_id).await?;
